Added changelog for TS, structured data and OG/twitter tags. (#15)
* added opengraph, twitter and structured data.

This is to make sure content can be found correctly, along with rendering in slack and twitter properly. We will need to get to static rendering, so we can correctly render this for each page.

Signed-off-by: Dave Shanley <dshanley@vmware.com>

* Added changelog component and rendering.

The changelog currently exists inside transport-typescript. This component extracts it and renders the data in a datagrid.

Signed-off-by: Dave Shanley <dshanley@vmware.com>